requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.28.1
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bleach==6.1.0
blinker==1.8.2
bokeh==3.4.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.6
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docarray==0.32.1
executing==2.0.1
experimental==0.9.7
faiss-cpu==1.8.0
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.1
graphviz==0.20.3
greenlet==3.0.3
groq==0.5.0
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
jsonpatch==1.33
jsonpickle==3.2.1
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_bokeh==4.0.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langchain
langchain-community
langchain-core
langchain-groq
langchain-openai
langchain-text-splitters
langchainhub
langgraph==0.0.49
langsmith==0.1.125
linkify-it-py==2.0.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
monotonic==1.6
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.46.1
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-alephalpha==0.23.0
opentelemetry-instrumentation-anthropic==0.23.0
opentelemetry-instrumentation-bedrock==0.23.0
opentelemetry-instrumentation-chromadb==0.23.0
opentelemetry-instrumentation-cohere==0.23.0
opentelemetry-instrumentation-google-generativeai==0.23.0
opentelemetry-instrumentation-haystack==0.23.0
opentelemetry-instrumentation-langchain==0.23.0
opentelemetry-instrumentation-llamaindex==0.23.0
opentelemetry-instrumentation-milvus==0.23.0
opentelemetry-instrumentation-mistralai==0.23.0
opentelemetry-instrumentation-ollama==0.23.0
opentelemetry-instrumentation-openai==0.23.0
opentelemetry-instrumentation-pinecone==0.23.0
opentelemetry-instrumentation-qdrant==0.23.0
opentelemetry-instrumentation-replicate==0.23.0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-sqlalchemy==0.46b0
opentelemetry-instrumentation-together==0.23.0
opentelemetry-instrumentation-transformers==0.23.0
opentelemetry-instrumentation-urllib3==0.46b0
opentelemetry-instrumentation-vertexai==0.23.0
opentelemetry-instrumentation-watsonx==0.23.0
opentelemetry-instrumentation-weaviate==0.23.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-semantic-conventions-ai==0.3.1
opentelemetry-util-http==0.46b0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
panel==1.4.4
param==2.1.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.1
posthog==3.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybind11==2.12.0
pydantic>=1.9.0
pydantic_core>=2.18.2
# QLoRA微调相关依赖
transformers>=4.40.0
accelerate>=0.31.0
peft>=0.11.0
bitsandbytes>=0.43.0
datasets>=2.19.0
torch>=2.4.0
scipy>=1.11.0
pydeck==0.9.1
pyee==11.1.0
Pygments==2.18.0
pylcs==0.1.1
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pyppeteer
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyvis==0.3.2
pyviz_comms==3.0.2
pywin32==306; platform_system == "Windows"
PyYAML==6.0.1
pyzmq==26.0.3
ragas==0.1.7
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
streamlit==1.32.0
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traceloop-sdk==0.23.0
traitlets==5.14.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
urllib3
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
widgetsnbextension==4.0.11
wrapt==1.16.0
xxhash==3.4.1
xyzservices==2024.6.0
yarl==1.9.4
zipp==3.18.2
